% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methods.R
\name{print.parboost}
\alias{print.parboost}
\title{Prints a short description of a parboost object.}
\usage{
\method{print}{parboost}(x, ...)
}
\arguments{
\item{x}{a parboost object.}

\item{...}{Additional arguements passed to callies.}
}
\value{
Prints a short descritpion of a parboost object.
}
\description{
Print a parboost object
}
\details{
Prints a basic description of a parboost object
}
\author{
Ronert Obst
}

