% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methods.R
\name{print.summary.parboost}
\alias{print.summary.parboost}
\title{Prints a summary of a parboost object.}
\usage{
\method{print}{summary.parboost}(x, ...)
}
\arguments{
\item{x}{a parboost object.}

\item{...}{Additional arguements passed to callies.}
}
\value{
Prints a summary of a parboost object.
}
\description{
Print a summary of a parboost object
}
\details{
Prints a basic summary of a parboost object
}
\author{
Ronert Obst
}

